Run rails app:update to complete the update to Rails 7.1 - Fix various compatibility issues with Rails 7.1 - Fix issue with zeitwerk mode in lib/errors.rb (renamed to lib/unparseable_response_error.rb) - Fix some deprecation warnings.
