+++ config/application.rb
@@ -4,6 +4,8 @@ require 'rails'
 # Pick the frameworks you want:
 require 'active_model/railtie'
 # require "active_job/railtie"
+
+# TODO: try to remove this:
 require 'active_record/attribute_assignment'
 # require "active_record/railtie"
 # require "active_storage/engine"
@@ -13,6 +15,7 @@ require 'action_mailer/railtie'
 # require "action_text/engine"
 require 'action_view/railtie'
 # require "action_cable/engine"
+# require "rails/test_unit/railtie"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -25,7 +28,6 @@ module TradeTariffFrontend
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 7.0
     config.time_zone = 'UTC'
-
     require 'trade_tariff_frontend'
 
     # Custom directories with classes and modules you want to be autoloadable.
@@ -65,5 +67,21 @@ module TradeTariffFrontend
       username == TradeTariffFrontend.basic_username &&
         password == TradeTariffFrontend.basic_password
     end
+
+    # Please, add to the `ignore` list any other `lib` subdirectories that do
+    # not contain `.rb` files, or that should not be reloaded or eager loaded.
+    # Common ones are `templates`, `generators`, or `middleware`, for example.
+    config.autoload_lib(ignore: %w[assets tasks])
+
+    # Configuration for the application, engines, and railties goes here.
+    #
+    # These settings can be overridden in specific environments using the files
+    # in config/environments, which are processed later.
+    #
+    # config.time_zone = "Central Time (US & Canada)"
+    # config.eager_load_paths << Rails.root.join("extras")
+
+    # Don't generate system test files.
+    config.generators.system_tests = nil
   end
 end
