+++ config/initializers/content_security_policy.rb
@@ -1,9 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-# Define an application-wide content security policy
-# For further information see the following documentation
-# https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy
-
+# Define an application-wide content security policy.
+# See the Securing Rails Applications Guide for more information:
+# https://guides.rubyonrails.org/security.html#content-security-policy-header
 Rails.application.configure do
   config.content_security_policy do |policy|
     policy.default_src :self
@@ -15,12 +14,4 @@ Rails.application.configure do
     # Specify URI for violation reports
     policy.report_uri ENV['SENTRY_CSP_ENDPOINT'] if ENV['SENTRY_CSP_ENDPOINT'].present?
   end
-  #
-  #   # Generate session nonces for permitted importmap and inline scripts
-  config.content_security_policy_nonce_generator = ->(request) { request.session.id.to_s }
-  #   config.content_security_policy_nonce_directives = %w(script-src)
-  #
-  #   Report CSP violations to a specified URI. See:
-  #   https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy-Report-Only
-  #   config.content_security_policy_report_only = true
 end
