+++ lib/api_entity.rb
@@ -1,7 +1,7 @@
 require 'multi_json'
 require 'active_model'
 require 'tariff_jsonapi_parser'
-require 'errors'
+require 'unparseable_response_error'
 
 module ApiEntity
   extend ActiveSupport::Concern
