+++ lib/unparseable_response_error.rb
